--- frontend/src/Logo.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-function Logo() {
-  return (
-    <>
-      <div id="title">
-        <h3>이화</h3>
-        <h3 id="green">공유배달</h3>
-      </div>
-    </>
-  );
-}
-
-export default Logo;
