+++ frontend/src/pages/login/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/named, import/prefer-default-export
+export * from './Login';
