+++ frontend/src/pages/main/Logo.js
@@ -1,0 +1,8 @@
+export default function Logo() {
+  return (
+    <div id="title">
+      <h3>이화</h3>
+      <h3 id="green">공유배달</h3>
+    </div>
+  );
+}
