+++ frontend/src/pages/main/Main.js
@@ -1,10 +1,9 @@
-import React from 'react';
-import './App.css';
+import '../../App.css';
 import Posts from './Posts';
 import FoodTypes from './FoodType';
 import Logo from './Logo';
 
-function MainScreen() {
+export default function Main() {
   return (
     <>
       <div id="topbar">
@@ -29,5 +28,3 @@ function MainScreen() {
     </>
   );
 }
-
-export default MainScreen;
