+++ frontend/src/pages/main/index.js
@@ -1,0 +1,2 @@
+export * from './Main';
+export * from './Logo';
