--- frontend/src/setProxy.js
@@ -1,11 +1,0 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-
-module.exports = function(app) {
-  app.use(
-    '/api',
-    createProxyMiddleware({
-      target: 'http://localhost:8080',	
-      changeOrigin: true,
-    })
-  );
-};
